requirements.txt
fastapi[standard]
websockets
python-dotenv
twilio
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
pandas
numpy
openai
celery
redis
